task-06/requirements.txt

python-dotenv>=1.0.0
discord.py>=2.3.1
beautifulsoup4>=4.12.2
bs4>=0.0.1
